Refactor completed tasks
- Move state down to CompletedTaskList

- Render CompletedTasks by id
